# m8flow-backend/src/m8flow_backend/routes/v1.py
from __future__ import annotations

from flask import Flask, g, jsonify, request

from m8flow_backend import catalog, human_task, identity, secrets, workflow
from m8flow_backend.auth import (
    authentication_identifier_for_request,
    clear_dead_auth_realm_cookie,
    encode_auth_token,
    on_login_or_token_enrichment,
    require_current_user,
    set_selected_tenant_cookie,
)
from m8flow_backend.authorization import allow_uri
from m8flow_backend.integrations.auth.keycloak.config import master_realm_name
from m8flow_backend.errors import ApiError
from m8flow_backend.routes import login_controller
from m8flow_backend.tenancy import SELECTED_TENANT_COOKIE_NAME, get_healthy_response, is_super_admin_request


def register_v1_routes(app: Flask) -> None:
    @app.get("/v1.0/status")
    @app.get("/v1.0/ping")
    @app.get("/v1.0/healthy")
    def status():
        payload, code = get_healthy_response()
        return jsonify(payload), code

    @app.get("/v1.0/onboarding")
    def onboarding():
        user = require_current_user()
        session = g.db_session
        if not allow_uri(user, "GET", "/v1.0/onboarding", session=session):
            raise ApiError("permission_denied", "Not allowed to read onboarding", 403)
        tenant_id = request.cookies.get(SELECTED_TENANT_COOKIE_NAME) or getattr(g, "m8flow_tenant_id", None)
        return jsonify({"ok": True, "username": user.username, "tenant_id": tenant_id})

    @app.get("/v1.0/tasks")
    def list_tasks():
        user = require_current_user()
        session = g.db_session
        if not allow_uri(user, "GET", "/v1.0/tasks", session=session):
            raise ApiError("permission_denied", "Not allowed to list tasks", 403)
        tenant_id = _require_tenant()
        if is_super_admin_request():
            tasks = workflow.list_pending_tasks_for_super_admin(session)
        else:
            tasks = workflow.list_pending_tasks(session, tenant_id=tenant_id, user_id=user.id)
        return jsonify(
            [
                {
                    "id": task.id,
                    "task_name": task.task_name,
                    "task_title": task.task_title,
                    "process_instance_id": task.process_instance_id,
                }
                for task in tasks
            ]
        )

    @app.put("/v1.0/tasks/<int:human_task_id>/claim")
    def claim_task(human_task_id: int):
        user = require_current_user()
        session = g.db_session
        tenant_id = _require_tenant()
        task = workflow.claim(session, tenant_id=tenant_id, human_task_id=human_task_id, user_id=user.id)
        return jsonify({"id": task.id, "actual_owner_id": task.actual_owner_id})

    @app.post("/v1.0/tasks/<int:human_task_id>/complete")
    def complete_task(human_task_id: int):
        user = require_current_user()
        session = g.db_session
        tenant_id = _require_tenant()
        payload = request.get_json(silent=True) or {}
        instance = workflow.complete(
            session,
            tenant_id=tenant_id,
            human_task_id=human_task_id,
            user_id=user.id,
            task_payload=payload,
        )
        return jsonify({"process_instance_id": instance.id, "status": instance.status})

    @app.get("/v1.0/tasks/<int:human_task_id>")
    def get_task(human_task_id: int):
        require_current_user()
        session = g.db_session
        tenant_id = _require_tenant()
        return jsonify(human_task.display_task(session, tenant_id=tenant_id, human_task_id=human_task_id))

    @app.post("/v1.0/process-models")
    def save_process_model():
        user = require_current_user()
        session = g.db_session
        tenant_id = _require_tenant()
        body = request.get_json(force=True)
        catalog.save(
            session,
            path=body["path"],
            xml=body["xml"],
            tenant_id=tenant_id,
            user_id=user.id,
        )
        return jsonify({"ok": True}), 201

    @app.get("/v1.0/process-models")
    def list_process_models():
        require_current_user()
        tenant_id = _require_tenant()
        group = request.args.get("group")
        return jsonify(catalog.list_models(group, tenant_id=tenant_id))

    @app.post("/v1.0/process-instances")
    def start_process():
        user = require_current_user()
        session = g.db_session
        tenant_id = _require_tenant()
        body = request.get_json(force=True)
        instance = workflow.start(
            session,
            tenant_id=tenant_id,
            user_id=user.id,
            process_model_identifier=body["process_model_identifier"],
        )
        return jsonify({"id": instance.id, "status": instance.status}), 201

    @app.get("/v1.0/process-instances")
    def list_instances():
        require_current_user()
        session = g.db_session
        tenant_id = _require_tenant()
        if is_super_admin_request():
            rows = workflow.list_instances_for_super_admin(session)
        else:
            rows = workflow.list_instances(session, tenant_id=tenant_id)
        return jsonify([{"id": row.id, "status": row.status} for row in rows])

    @app.get("/v1.0/secrets")
    def list_secrets():
        require_current_user()
        session = g.db_session
        tenant_id = _require_tenant()
        from sqlalchemy import select
        from m8flow_backend.models.native import SecretModel

        rows = session.scalars(select(SecretModel).where(SecretModel.m8f_tenant_id == tenant_id)).all()
        return jsonify([{"key": row.key} for row in rows])

    @app.put("/v1.0/secrets/<key>")
    def put_secret(key: str):
        require_current_user()
        session = g.db_session
        tenant_id = _require_tenant()
        body = request.get_json(force=True)
        secrets.put_secret(session, tenant_id=tenant_id, key=key, value=body["value"])
        return jsonify({"key": key})

    @app.post("/v1.0/login")
    def login():
        session = g.db_session
        body = request.get_json(force=True)
        tenant_id = body.get("tenant_id") or request.cookies.get(SELECTED_TENANT_COOKIE_NAME)
        if not tenant_id:
            raise ApiError("tenant_required", "tenant_id is required", 400)
        user = on_login_or_token_enrichment(
            session,
            username=body["username"],
            service=body.get("service") or "local",
            service_id=body.get("service_id") or body["username"],
            email=body.get("email"),
            active_tenant_id=tenant_id,
        )
        groups = body.get("groups") or []
        identity.sync_groups(session, user=user, group_identifiers=groups, tenant_id=tenant_id)
        token = encode_auth_token(user=user)
        response = jsonify({"access_token": token, "token_type": "Bearer"})
        set_selected_tenant_cookie(response, tenant_id)
        clear_dead_auth_realm_cookie(response)
        _ = authentication_identifier_for_request()
        _ = master_realm_name()
        return response

    # Browser-redirect Keycloak login/logout (distinct from the JSON POST /v1.0/login
    # above): a full-page GET on the same path, dispatched separately by method.
    # See routes/login_controller.py.
    app.add_url_rule(
        "/v1.0/login",
        endpoint="browser_login",
        view_func=login_controller.login,
        methods=["GET"],
    )
    app.add_url_rule(
        "/v1.0/login_return",
        endpoint="login_return",
        view_func=login_controller.login_return,
        methods=["GET"],
    )
    app.add_url_rule(
        "/v1.0/logout",
        endpoint="logout",
        view_func=login_controller.logout,
        methods=["GET"],
    )
    app.add_url_rule(
        "/v1.0/refresh",
        endpoint="refresh_tokens",
        view_func=login_controller.refresh,
        methods=["POST"],
    )

    @app.post("/v1.0/m8flow/external-forms/<int:human_task_id>/submit")
    def submit_external_form(human_task_id: int):
        user = require_current_user()
        session = g.db_session
        tenant_id = _require_tenant()
        payload = request.get_json(silent=True) or {}
        instance = human_task.submit_external_form(
            session,
            tenant_id=tenant_id,
            human_task_id=human_task_id,
            user_id=user.id,
            task_payload=payload,
        )
        return jsonify({"process_instance_id": instance.id, "status": instance.status})


def _require_tenant() -> str:
    tenant_id = request.cookies.get(SELECTED_TENANT_COOKIE_NAME) or getattr(g, "m8flow_tenant_id", None)
    if not tenant_id:
        raise ApiError("tenant_required", "m8flow_selected_tenant cookie is required", 400)
    g.m8flow_tenant_id = tenant_id
    return tenant_id